+++ gettsim/tax_transfer.py
@@ -9,7 +9,7 @@ from gettsim.pensions import pensions
 from gettsim.pre_processing.apply_tax_funcs import apply_tax_transfer_func
 from gettsim.pre_processing.checks import check_boolean
 from gettsim.pre_processing.policy_for_date import get_policies_for_date
-from gettsim.social_insurance import soc_ins_contrib
+from gettsim.soz_vers import soc_ins_contrib
 from gettsim.taxes.eink_st import eink_st
 from gettsim.taxes.favorability_check import favorability_check
 from gettsim.taxes.kindergeld import kindergeld
