+++ gettsim/tests/test_soc_ins_contrib.py
@@ -6,7 +6,7 @@ import pytest
 from gettsim.config import ROOT_DIR
 from gettsim.pre_processing.apply_tax_funcs import apply_tax_transfer_func
 from gettsim.pre_processing.policy_for_date import get_policies_for_date
-from gettsim.social_insurance import soc_ins_contrib
+from gettsim.soz_vers import soc_ins_contrib
 
 INPUT_COLS = [
     "p_id",
